+++ cudalib/Cuda/internal.h
@@ -225,7 +225,7 @@ void calcCorr(const Mat33& Rcurr,
 	float angleThres,
 	DeviceArray<jtj> & sum,
 	DeviceArray<jtj> & out,
-	float * matrixA_host,
+	double * matrixA_host,
 	int * result,
 	int threads, int blocks);
